Render pong in the default index page
