+++ pages/index.tsx
@@ -1,63 +1,23 @@
+import React from "react";
 import type { NextPage } from "next";
-import Head from "next/head";
+import dynamic from "next/dynamic";
 
-import Counter from "../src/features/counter/Counter";
-import styles from "../src/styles/Home.module.css";
+import Instructions from "../src/components/Instructions";
 
-const IndexPage: NextPage = () => {
+const PongContainerWithNoSSR = dynamic(
+  () => import("../src/features/pong/PongContainer"),
+  {
+    ssr: false,
+  }
+);
+
+export const PongApp: NextPage = (props) => {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Redux Toolkit</title>
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <header className={styles.header}>
-        <img src="/logo.svg" className={styles.logo} alt="logo" />
-        <Counter />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <span>
-          <span>Learn </span>
-          <a
-            className={styles.link}
-            href="https://reactjs.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            React
-          </a>
-          <span>, </span>
-          <a
-            className={styles.link}
-            href="https://redux.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Redux
-          </a>
-          <span>, </span>
-          <a
-            className={styles.link}
-            href="https://redux-toolkit.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Redux Toolkit
-          </a>
-          ,<span> and </span>
-          <a
-            className={styles.link}
-            href="https://react-redux.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            React Redux
-          </a>
-        </span>
-      </header>
+    <div className="appContainer">
+      <PongContainerWithNoSSR />
+      <Instructions />
     </div>
   );
 };
 
-export default IndexPage;
+export default PongApp;
